requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
commonmark==0.9.1
dnspython==2.2.1
idna==3.3
Naked==0.1.31
numpy==1.23.0
pandas==1.4.3
pycryptodome==3.15.0
Pygments==2.12.0
pymongo==4.1.1
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rich==12.4.4
shellescape==3.8.1
six==1.16.0
urllib3==1.26.9
